// src/app/models/request.ts
import { Instance } from '../models/platform';

export class IdolRequest{
    Text:string;
    Server: string;
    Port: string;
    DataBase: string;
    FieldText: string;
    PrintFields: string;
    Sort: string;
    Summary: string;
    Instance: Instance;
    Environment: string;
    UrlExists: string;
    Print: string;
    Start: string;
    End: string;
}